requirements.txt
discord.py>=2.4.0
yt-dlp
python-dotenv   # optional